requirements.txt
anyio==4.4.0
asgiref==3.8.1
beautifulsoup4==4.12.3
Cerberus==1.3.5
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==5.0.6
django-environ==0.11.2
django-filter==24.2
djangorestframework==3.15.1
extruct==0.17.0
h11==0.14.0
html5lib==1.1
html_text==0.6.2
httpcore==1.0.5
httpx==0.27.0
idna==3.7
isodate==0.6.1
Jinja2==3.1.4
jstyleson==0.0.2
kolo==2.27.0
lxml==5.2.2
lxml_html_clean==0.1.1
Markdown==3.6
MarkupSafe==2.1.5
mf2py==2.0.1
more-itertools==10.2.0
msgpack==1.0.8
numpy==2.0.0
pandas==2.2.2
pillow==10.3.0
psycopg2-binary==2.9.9
pyparsing==3.1.2
pyRdfa3==3.6.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
rdflib==7.0.0
recipe_scrapers==14.56.0
requests==2.32.2
setuptools==70.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlglot==24.1.2
sqlparse==0.5.0
tabulate==0.9.0
toolz==0.12.1
tzdata==2024.1
ulid-py==1.1.0
urllib3==2.2.1
w3lib==2.1.2
webencodings==0.5.1
wheel==0.43.0
whitenoise==6.6.0
